After Merging Suresh branch, all files checked.

+++ src/main/java/com/callippus/water/erp/domain/BillAndCollectionDTO.java
@@ -5,9 +5,9 @@ import java.time.ZonedDateTime;
 public class BillAndCollectionDTO {
 	ZonedDateTime txn_date;
 	String txn_type;
-	Float debit;
-	Float credit;
-	Float balance;
+	Double debit;
+	Double credit;
+	Double balance;
 	public ZonedDateTime getTxn_date() {
 		return txn_date;
 	}
@@ -20,29 +20,30 @@ public class BillAndCollectionDTO {
 	public void setTxn_type(String txn_type) {
 		this.txn_type = txn_type;
 	}
-	public Float getDebit() {
+	public Double getDebit() {
 		return debit;
 	}
-	public void setDebit(Float debit) {
+	public void setDebit(Double debit) {
 		this.debit = debit;
 	}
-	public Float getCredit() {
+	public Double getCredit() {
 		return credit;
 	}
-	public void setCredit(Float credit) {
+	public void setCredit(Double credit) {
 		this.credit = credit;
 	}
-	public Float getBalance() {
+	public Double getBalance() {
 		return balance;
 	}
-	public void setBalance(Float balance) {
+	public void setBalance(Double balance) {
 		this.balance = balance;
 	}
+	
 	@Override
 	public String toString() {
 		return "BillAndCollectionDTO [txn_date=" + txn_date + ", txn_type=" + txn_type + ", debit=" + debit
 				+ ", credit=" + credit + ", balance=" + balance + "]";
 	}
-	
+
 	
 }
